feat: workorder garden and garden utils package
