--- src/apps/WorkOrder/Garden/components/Icons/index.ts
@@ -1,3 +1,0 @@
-export * from './StatusCircle';
-export * from './FlagIcon';
-export * from './SizeIcons';
