+++ src/apps/WorkOrder/Garden/components/Popover/styles.ts
@@ -1,44 +1,5 @@
 import styled from 'styled-components';
 
-export const PopoverContainer = styled.div`
-    white-space: nowrap;
-
-    hr {
-        border: 1px solid #dcdcdc;
-    }
-
-    h5 {
-        margin-top: 0;
-        margin-bottom: 0;
-    }
-
-    p {
-        margin: 0;
-        font-size: 12px;
-    }
-`;
-type CommStatusProps = {
-    barColor: string;
-    textColor: string;
-};
-export const CommStatus = styled.div<CommStatusProps>`
-    display: flex;
-    border-radius: 4px;
-    font-size: 12px;
-    height: 24px;
-    padding: 4px 8px;
-    margin-top: 16px;
-    text-transform: capitalize;
-    background: ${(p) => p.barColor};
-    color: ${(p) => p.textColor};
-    line-height: 24px;
-    text-align: center;
-    justify-content: space-between;
-    > strong:first-child {
-        margin-right: 32px;
-    }
-`;
-
 export const Statuses = styled.div`
     margin-top: 15px;
     display: flex;
@@ -54,22 +15,7 @@ export const Statuses = styled.div`
         }
     }
 `;
-type StatusProps = {
-    color: string;
-};
-export const Status = styled.div<StatusProps>`
-    width: 40px;
-    height: 24px;
-    display: flex;
-    align-self: center;
-    border: none;
-    align-items: center;
-    justify-content: center;
-    font-size: 12px;
-    font-weight: 600;
-    border-radius: 12px;
-    background: ${(p) => p.color};
-`;
+
 export const HoldBy = styled.div`
     display: flex;
     width: 100%;
