+++ src/apps/WorkOrder/Garden/components/WorkOrderSidesheet/Tabs/DetailsTab.tsx
@@ -1,8 +1,7 @@
-import { tokens } from '@equinor/eds-tokens';
+import { formatDateString, materialStatusMap, StringCell, Table } from '@equinor/GardenUtils';
 import styled from 'styled-components';
 import { WorkOrder } from '../../../models';
-import { formatDateString, getWoStatus, materialStatusMap } from '../../../utility';
-const StringCell = ({ value }: { value: string }) => <>{value ? value.trim() : 'N/A'}</>;
+import { getWoStatus } from '../../../utility';
 
 const TabContent = styled.div`
     height: 100%;
@@ -12,26 +11,7 @@ const TabContent = styled.div`
         padding: 8px;
     }
 `;
-export const Table = styled.table`
-    width: 100%;
-    border-spacing: 0;
-    border-collapse: collapse;
-    margin-bottom: 32px;
 
-    th {
-        text-align: left;
-    }
-
-    td {
-        border-bottom: 1px solid ${tokens.colors.ui.background__medium.hex};
-        padding: 8px;
-
-        &:first-child {
-            width: 240px;
-            font-weight: bold;
-        }
-    }
-`;
 type DetailsTabProps = {
     workOrder: WorkOrder;
 };
