+++ src/apps/WorkOrder/Garden/models/index.ts
@@ -1,6 +1,3 @@
 export * from './workOrder';
-export * from './pcsStatus';
-export * from './pcsFollowUpStatus';
-export * from './materialStatus';
 export * from './workOrderMaterial';
 export * from './workOrderMccr';
