--- src/apps/WorkOrder/Garden/models/pcsFollowUpStatus.ts
@@ -1,5 +1,0 @@
-export type PCSFollowUpStatus =
-    | 'WOFinished'
-    | 'MaterialAndWoOk'
-    | 'MaterialAndWoAvailable'
-    | 'MaterialAndOrWoNotAvailable';
