--- src/apps/WorkOrder/Garden/models/pcsStatus.ts
@@ -1,12 +1,0 @@
-export type PCSStatus =
-    | 'NoStatus'
-    | 'Prepared'
-    | 'Cancelled'
-    | 'ToMC'
-    | 'MCDocsPrepared'
-    | 'ToField'
-    | 'FromField'
-    | 'ComplByMC'
-    | 'SentDC'
-    | 'SentToPlanning'
-    | 'ASBuiltCompleted';
