+++ src/apps/WorkOrder/Garden/utility/pcsWorkOrder.ts
@@ -1,31 +1,5 @@
-import { PCSStatus } from '../models';
+import { ProcosysStatuses } from '@equinor/GardenUtils';
 
-const proCoSysStatuses: Record<PCSStatus, string> = {
-    NoStatus: 'No status',
-    Prepared: 'WO Prepared',
-    Cancelled: 'WO Cancelled',
-    ToMC: 'WO to MC',
-    MCDocsPrepared: "MC doc's prepared",
-    ToField: 'WO to Field',
-    FromField: 'WO from Field',
-    ComplByMC: 'WO compl. By MC',
-    SentDC: 'WO Sent DC',
-    SentToPlanning: 'WO Sent to planning',
-    ASBuiltCompleted: 'AS Built completed',
-};
-export enum ProcosysStatuses {
-    NoStatus = 'No status',
-    Prepared = 'WO Prepared',
-    Cancelled = 'WO Cancelled',
-    ToMC = 'WO to MC',
-    MCDocsPrepared = "MC doc's prepared",
-    ToField = 'WO to Field',
-    FromField = 'WO from Field',
-    ComplByMC = 'WO compl. By MC',
-    SentDC = 'WO Sent DC',
-    SentToPlanning = 'WO Sent to planning',
-    ASBuiltCompleted = 'AS Built completed',
-}
 export const proCoSysStatusPriorityMap: Record<ProcosysStatuses, number> = {
     [ProcosysStatuses.ASBuiltCompleted]: 0,
     [ProcosysStatuses.SentToPlanning]: 1,
