+++ src/packages/GardenUtils/src/components/popover/PopoverContainer.ts
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const PopoverContainer = styled.div`
+    white-space: nowrap;
+
+    hr {
+        border: 1px solid #dcdcdc;
+    }
+
+    h5 {
+        margin-top: 0;
+        margin-bottom: 0;
+    }
+
+    p {
+        margin: 0;
+        font-size: 12px;
+    }
+`;
