+++ src/packages/GardenUtils/src/components/popover/PopoverStatus.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+type StatusProps = {
+    color: string;
+};
+export const PopoverStatus = styled.div<StatusProps>`
+    width: 40px;
+    height: 24px;
+    display: flex;
+    align-self: center;
+    border: none;
+    align-items: center;
+    justify-content: center;
+    font-size: 12px;
+    font-weight: 600;
+    border-radius: 12px;
+    background: ${(p) => p.color};
+`;
