+++ src/packages/GardenUtils/src/components/sidesheet/SidesheetContainer.ts
@@ -1,9 +1,13 @@
 import styled from 'styled-components';
-
+/**
+ * Styled component to wrap the package's sidesheet in.
+ */
 export const SideSheetContainer = styled.div`
     display: flex;
     flex-direction: column;
     box-sizing: border-box;
     padding: 16px;
     width: 100%;
+    height: 100%;
+    overflow: auto;
 `;
