+++ src/packages/GardenUtils/src/mapping/index.ts
@@ -1,0 +1,1 @@
+export * from './materialStatusMap';
