+++ src/packages/GardenUtils/src/mapping/materialStatusMap.ts
@@ -1,0 +1,16 @@
+import { MaterialStatus } from '..';
+/**
+ * Dictionary for mapping material statuses to readable string.
+ */
+export const materialStatusMap: Partial<Record<MaterialStatus, string>> = {
+    M10: 'Material requested to job site',
+    M12: 'Material received on job site',
+    M2: 'Materials linked to Smartpack/Jobcard',
+    M6: 'Material partly delivered',
+    M7: 'Materials fully delivered',
+    M9: 'Material returned',
+    MN: 'No Material required',
+    MN1: 'Additional material to be issued Offshore from Min/Max Stock',
+    MNX1: 'Materials not linked to Smartpack/Jobcard',
+    MNX2: 'Materials partially linked to Smartpack/Jobcard',
+};
