+++ src/packages/GardenUtils/src/models/pcsFollowUpStatus.ts
@@ -1,0 +1,6 @@
+export enum FollowUpStatuses {
+    WOFinished = 'WO Finished',
+    MaterialAndWoOk = 'Material and WorkOrder OK',
+    MaterialAndWoAvailable = 'Material and WorkOrder Available',
+    MaterialAndOrWoNotAvailable = 'Material and/or WorkOrder not Available',
+}
