+++ src/packages/GardenUtils/src/models/pcsStatus.ts
@@ -1,0 +1,13 @@
+export enum ProcosysStatuses {
+    NoStatus = 'No status',
+    Prepared = 'WO Prepared',
+    Cancelled = 'WO Cancelled',
+    ToMC = 'WO to MC',
+    MCDocsPrepared = "MC doc's prepared",
+    ToField = 'WO to Field',
+    FromField = 'WO from Field',
+    ComplByMC = 'WO compl. By MC',
+    SentDC = 'WO Sent DC',
+    SentToPlanning = 'WO Sent to planning',
+    ASBuiltCompleted = 'AS Built completed',
+}
