+++ src/packages/GardenUtils/src/sortUtils.ts
@@ -1,0 +1,2 @@
+export const sortByNumber = (a: string, b: string): number =>
+    a.localeCompare(b, undefined, { numeric: true, sensitivity: 'base' });
