+++ src/packages/GardenUtils/src/statusUtils.ts
@@ -1,0 +1,7 @@
+type Status = 'OS' | 'PB' | 'PA' | 'OK';
+export const statusColorMap: Record<Status, string> = {
+    OS: '#9e9e9e',
+    PB: '#ffc107',
+    PA: '#ff4081',
+    OK: '#00c853',
+};
